wip refactor DiscoverResult
