+++ bundles/org.openhab.binding.souliss/src/main/java/org/openhab/binding/souliss/internal/discovery/SoulissDiscoverJob.java
@@ -12,7 +12,6 @@
  */
 package org.openhab.binding.souliss.internal.discovery;
 
-import java.net.InetAddress;
 import java.util.Collection;
 import java.util.concurrent.ConcurrentMap;
 
@@ -33,19 +32,6 @@ import org.slf4j.LoggerFactory;
 @NonNullByDefault
 public class SoulissDiscoverJob implements Runnable {
     private CommonCommands soulissCommands = new CommonCommands();
-
-    /**
-     * Result callback interface.
-     */
-    public interface DiscoverResult {
-        static boolean IS_GATEWAY_DETECTED = false;
-
-        void gatewayDetected(InetAddress addr, String id);
-
-        void thingDetectedTypicals(byte lastByteGatewayIP, byte typical, byte node, byte slot);
-
-        void thingDetectedActionMessages(String sTopicNumber, String sTopicVariant);
-    }
 
     @Nullable
     UDPListenDiscoverRunnable udpServerOnDefaultPort = null;
