+++ bundles/org.openhab.binding.souliss/src/main/java/org/openhab/binding/souliss/internal/protocol/UDPListenDiscoverRunnable.java
@@ -22,7 +22,7 @@ import java.nio.channels.DatagramChannel;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.eclipse.jdt.annotation.Nullable;
-import org.openhab.binding.souliss.internal.discovery.SoulissDiscoverJob.DiscoverResult;
+import org.openhab.binding.souliss.internal.discovery.DiscoverResult;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
